+++ sdks/web-ui-sdk/CHANGELOG.md
@@ -1,5 +1,19 @@
 # web-ui-sdk
 
+## 1.5.71
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.70
+
+## 1.5.70
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.69
+
 ## 1.5.69
 
 ### Patch Changes