+++ sdks/workflow-browser-sdk/CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Updated dependencies
   - @ballerine/workflow-core@0.6.89
+  - @ballerine/common@0.9.70
 
 ## 0.6.88
 
@@ -13,6 +14,7 @@
 
 - Updated dependencies
   - @ballerine/workflow-core@0.6.88
+  - @ballerine/common@0.9.69
 
 ## 0.6.87