+++ services/workflows-service/CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Updated dependencies
   - @ballerine/workflow-core@0.6.89
+  - @ballerine/common@0.9.70
   - @ballerine/workflow-node-sdk@0.6.89
 
 ## 0.7.92
@@ -14,6 +15,7 @@
 
 - Updated dependencies
   - @ballerine/workflow-core@0.6.88
+  - @ballerine/common@0.9.69
   - @ballerine/workflow-node-sdk@0.6.88
 
 ## 0.7.91